refactor(ai): lift RuleFlag into @carebridge/shared-types and type rules_output
Move the RuleFlag interface from ai-oversight/rules/critical-values.ts
into packages/shared-types/src/ai-flags.ts so downstream consumers get
compile-time guarantees on the stored shape. Update the Drizzle schema
to use RuleFlag[] for review_jobs.rules_output, eliminating the double
cast (as unknown as Array<Record<string, unknown>>) in review-service.

Closes #535
