+++ services/ai-oversight/src/rules/allergy-medication.ts
@@ -9,8 +9,7 @@
  * Rule IDs follow pattern: ALLERGY-MED-{SEQ}
  */
 
-import type { FlagSeverity, FlagCategory } from "@carebridge/shared-types";
-import type { RuleFlag } from "./critical-values.js";
+import type { FlagSeverity, FlagCategory, RuleFlag } from "@carebridge/shared-types";
 import type { PatientContext } from "./cross-specialty.js";
 
 /**
