+++ services/ai-oversight/src/rules/message-screening.ts
@@ -9,8 +9,7 @@
  * Does NOT auto-reply to patients — only creates flags for clinical review.
  */
 
-import type { FlagCategory, ClinicalEvent } from "@carebridge/shared-types";
-import type { RuleFlag } from "./critical-values.js";
+import type { FlagCategory, ClinicalEvent, RuleFlag } from "@carebridge/shared-types";
 import {
   SHARED_CRITICAL_PATTERNS,
   materializePattern,
